refactor: move LA Section 251 processing from ancillary
Local Authority Section 251 data are considered a "core" dataset.
