+++ data-pipeline/tests/unit/pre_processing/local_authorities/test_la_budget.py
@@ -3,7 +3,7 @@ import io
 import pandas as pd
 
 from pipeline import input_schemas
-from pipeline.pre_processing.ancillary import local_authority
+from pipeline.pre_processing import local_authority
 
 
 def test_la_budget(la_budget: pd.DataFrame):
