+++ data-pipeline/tests/unit/pre_processing/local_authorities/test_local_authorities.py
@@ -2,7 +2,7 @@ import io
 
 import pandas as pd
 
-from pipeline.pre_processing.ancillary import local_authority
+from pipeline.pre_processing import local_authority
 
 
 def test_local_authorities(
